refactor: renamae parser -> armtemplateparser
