+++ sg/internal/engine/builder.go
@@ -8,9 +8,9 @@ import (
 
 // QueryerBuilder constructs a Queryer.
 type QueryerBuilder struct {
-	packages   []policy.Package
-	queryCache QueryCache
-	err        error
+	packages                 []policy.Package
+	queryCache               QueryCache
+	err                      error
 	parseArmTemplateDefaults bool
 }
 
@@ -58,8 +58,8 @@ func (qb *QueryerBuilder) Complete() (Queryer, error) {
 		// NOTE: we limit the actual query by CPU count as policy evaluation is CPU bounded.
 		//       For input actions like reading policy files / source code, we allow them to run unbounded,
 		//       as the actual limiting is done by this limiter.
-		limiter:    newLimiterFromMaxProcs(),
-		queryCache: qb.queryCache,
+		limiter:                  newLimiterFromMaxProcs(),
+		queryCache:               qb.queryCache,
 		parseArmTemplateDefaults: qb.parseArmTemplateDefaults,
 	}
 	return rv, nil
